+++ packages/js-runtime/src/runtime/http/FormData.ts
@@ -1,9 +1,9 @@
 (globalThis => {
   // Almost the same implementation as URLSearchParams
   globalThis.FormData = class {
-    private fields: Map<string, string[]> = new Map();
+    private fields: Map<string, FormDataEntryValue[]> = new Map();
 
-    private addValue(name: string, value: string) {
+    private addValue(name: string, value: FormDataEntryValue) {
       const values = this.fields.get(name);
 
       if (values) {
@@ -13,15 +13,21 @@
       }
     }
 
-    append(name: string, value: string) {
-      this.addValue(name, value);
+    append(name: string, value: Blob, filename?: string): void;
+    append(name: string, value: string): void;
+    append(name: string, value: string | Blob, filename?: string) {
+      if (value instanceof Blob) {
+        this.addValue(name, new File([value], filename ?? ''));
+      } else {
+        this.addValue(name, value);
+      }
     }
 
     delete(name: string) {
       this.fields.delete(name);
     }
 
-    *entries(): IterableIterator<[string, string]> {
+    *entries(): IterableIterator<[string, FormDataEntryValue]> {
       for (const [key, values] of this.fields) {
         for (const value of values) {
           yield [key, value];
@@ -29,7 +35,7 @@
       }
     }
 
-    forEach(callbackfn: (value: string, key: string, parent: FormData) => void, thisArg?: any) {
+    forEach(callbackfn: (value: FormDataEntryValue, key: string, parent: FormData) => void, thisArg?: any) {
       this.fields.forEach((values, key) => {
         values.forEach(value => {
           callbackfn.call(thisArg, value, key, this);
@@ -37,11 +43,11 @@
       });
     }
 
-    get(name: string): string | null {
+    get(name: string): FormDataEntryValue | null {
       return this.fields.get(name)?.[0] || null;
     }
 
-    getAll(name: string): string[] {
+    getAll(name: string): FormDataEntryValue[] {
       return this.fields.get(name) || [];
     }
 
@@ -53,11 +59,17 @@
       return this.fields.keys();
     }
 
-    set(name: string, value: string) {
-      this.fields.set(name, [value]);
+    set(name: string, value: Blob, filename?: string): void;
+    set(name: string, value: string): void;
+    set(name: string, value: string | Blob, filename?: string) {
+      if (value instanceof Blob) {
+        this.fields.set(name, [new File([value], filename ?? '')]);
+      } else {
+        this.fields.set(name, [value]);
+      }
     }
 
-    *values(): IterableIterator<string> {
+    *values(): IterableIterator<FormDataEntryValue> {
       for (const [, values] of this.fields) {
         for (const value of values) {
           yield value;
@@ -65,7 +77,7 @@
       }
     }
 
-    [Symbol.iterator](): IterableIterator<[string, string]> {
+    [Symbol.iterator](): IterableIterator<[string, FormDataEntryValue]> {
       return this.entries();
     }
   };