+++ packages/ui/.storybook/main.ts
@@ -1,26 +1,29 @@
-const postcss = require('postcss');
-const tailwindcss = require('../tailwind.config');
-module.exports = {
+import postcss from 'postcss';
+import * as path from 'node:path';
+
+export default {
   stories: ['../src/components/**/*.stories.mdx', '../src/components/**/*.stories.@(js|jsx|ts|tsx)'],
-  addons: ['@storybook/addon-links', '@storybook/addon-essentials', '@storybook/addon-interactions', 'storybook-addon-next-router', {
-    name: '@storybook/addon-postcss',
-    options: {
-      postcssLoaderOptions: {
-        implementation: postcss,
-        postcssOptions: {
-          plugins: {
-            tailwindcss,
-            autoprefixer: {}
-          }
-        }
-      }
-    }
-  }],
+  addons: [
+    '@storybook/addon-links',
+    '@storybook/addon-essentials',
+    '@storybook/addon-interactions',
+    '@storybook/addon-styling',
+    {
+      name: '@storybook/addon-styling',
+      options: {
+        postCss: {
+          implementation: postcss,
+        },
+      },
+    },
+  ],
   framework: {
-    name: '@storybook/react-webpack5',
-    options: {}
+    name: '@storybook/nextjs',
+    options: {
+      nextConfigPath: path.resolve(__dirname, '../../dashboard/next.config.js'),
+    },
   },
   docs: {
-    autodocs: true
-  }
+    autodocs: true,
+  },
 };